Add support for importing LDR as a component and scaling physically.
Set the MAKE_COMPONENT config item to true and the imported LDR file will be
placed as a component definition.

If the above is true then setting PHYSICAL_SCALE to true will also scale the
component instance so it is in real world units, IE: 8mm between the centre
of studs.

Note that the scale is only applied to the top level component instance or
the instance of an imported individual part. The underlying part definitions
are still in LDR units.