add part plugin system
users can register a plugin that is called on part import to supply custom
part definitions.

there are 2 phases to this

1. When importing PartPlugin::import is called giving the plugin the chance to
import a custom definition for a part.

2. After import PartPlugin::post_import is called giving the plugin the chance
to modify an imported part. This is called for all parts whether they are
sourced from a plugin, SKP file or DAT file.

Note that are called for parts only, not for models.